remove the code that adds the vpn flag for clients
the developer of the ffmap software explained me that the vpn flag
should be set only for the vpn gateway server and all links to this
server are then considered as vpn links. If two clients are connected
through a mesh connection the link between them can be incorrectly considered as vpn
link if the vpn flag is set.